requirements.txt
flask
discord
openai
python-dotenv